+++ packages/core/src/data/index.ts
@@ -2,3 +2,5 @@ export * from "./casino";
 export * from "./chains";
 export * from "./tokens";
 export * from "./subgraphs/protocol";
+export * from "./api/freebet/campaigns";
+export * from "./api/freebet/bets";